fix[build]: 修复构建脚本和日志级别显示问题
1. 修改构建脚本，移除自动依赖收集的路径参数
2. 修复 SPDLOG 日志级别名称定义，使用 string_view 避免潜在问题
3. 修正 RGB 转换中的位操作类型转换问题
4. 更新项目规则文档，添加编译和注释语言要求
5. 更新项目待办事项，标记 invertAlpha 选项问题

+++ YUViewLib/src/video/rgb/ConversionRGB.cpp
@@ -621,10 +621,10 @@ void convertBitPackedToARGB(const QByteArray     &sourceBuffer,
                         : maxValue;
         if (bitOffset + bps > 32)
         {
-          valueR |= planeR[byteStart + 4] << 32;
-          valueG |= planeG[byteStart + 4] << 32;
-          valueB |= planeB[byteStart + 4] << 32;
-          valueA |= planeA ? planeA[byteStart + 4] << 32 : 0;
+          valueR |= static_cast<uint64_t>(planeR[byteStart + 4]) << 32;
+          valueG |= static_cast<uint64_t>(planeG[byteStart + 4]) << 32;
+          valueB |= static_cast<uint64_t>(planeB[byteStart + 4]) << 32;
+          valueA |= planeA ? static_cast<uint64_t>(planeA[byteStart + 4]) << 32 : 0;
         }
       }
     }
@@ -658,7 +658,7 @@ void convertBitPackedToARGB(const QByteArray     &sourceBuffer,
         value = rawData[byteStart] | (rawData[byteStart + 1] << 8) |
                 (rawData[byteStart + 2] << 16) | (rawData[byteStart + 3] << 24);
         if (bitOffset + bps > 32)
-          value |= rawData[byteStart + 4] << 32;
+          value |= static_cast<uint64_t>(rawData[byteStart + 4]) << 32;
       }
       valueR = (value >> rBitPos) & maxValue;
       valueG = (value >> gBitPos) & maxValue;
